requirements.txt
beautifulsoup4>=4.13.5
curl-cffi>=0.13.0
feedparser>=6.0.0
Flask>=3.1.2
google-adk>=1.21.0
haxor>=1.2.4
markitdown[pdf]>=0.1.3
requests>=2.32.5
